Add program to compare pairs of apertures and make plots
compare-pairs.py - Command line arguments to specify emission line,
spectral range, and data column.  Makes a separate graph for each pair
of apertures, showing the correlation between values of the data
column for overlapping apertures. Calculates the median ratio and
shows it as a line on the graph (green), together with the 1:1 line
(red dashed).